+++ gatsby-browser.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ContextProviderWithReducer } from './src/context/Context';
+import { ContextProviderWithReducer } from './src/context/Overlay';
 
 export const wrapRootElement = ({ element }) => {
   return <ContextProviderWithReducer>{element}</ContextProviderWithReducer>;
