+++ gatsby-ssr.js
@@ -1,11 +1,15 @@
 import React from 'react';
 import { renderToString } from 'react-dom/server';
-import { ContextProviderWithReducer } from './src/context/Context';
+
+import i18n from './i18n/config';
+import { ContextProviderWithReducer } from './src/context/Overlay';
 
 export const replaceRenderer = ({ bodyComponent, replaceBodyHTMLString }) => {
-  const App = () => (
-    <ContextProviderWithReducer>{bodyComponent}</ContextProviderWithReducer>
-  );
+  i18n.loadNamespaces(['translation'], () => {
+    const App = () => (
+      <ContextProviderWithReducer>{bodyComponent}</ContextProviderWithReducer>
+    );
 
-  replaceBodyHTMLString(renderToString(<App />));
+    replaceBodyHTMLString(renderToString(<App />));
+  });
 };
