+++ i18n/config.js
@@ -1,5 +1,4 @@
 import i18 from 'i18next';
-import { initReactI18next } from 'react-i18next';
 
 // i18next-node-fs-backend is not used because this file is imported and transpiled through Webpack.
 import resources from './resources';
@@ -7,7 +6,7 @@ import resources from './resources';
 import { defaultLangKey } from '../languages';
 
 // @see https://www.i18next.com/overview/configuration-options#configuration-options
-i18.use(initReactI18next).init({
+i18.init({
   debug: process.env.NODE_ENV === 'development',
   resources,
   fallbackLng: defaultLangKey,
