+++ src/components/LanguageList/LanguageListItem.jsx
@@ -6,7 +6,7 @@ import Link from '../Link/LinkGatsby';
 
 import Context, {
   SET_LANGUAGE_OVERLAY_VISIBILITY,
-} from '../../context/Context';
+} from '../../context/Overlay';
 
 // As close as possible to https://github.com/ec-europa/europa-component-library/blob/v2-dev/src/systems/ec/implementations/react/page-structure/language-list/src/LanguageListItem.jsx
 const LanguageListItem = ({ label, lang, href, isActive, isOverlay }) => {
