+++ src/components/LanguageList/LanguageListOverlayWithContext.jsx
@@ -2,7 +2,7 @@ import React, { useContext } from 'react';
 
 import Context, {
   SET_LANGUAGE_OVERLAY_VISIBILITY,
-} from '../../context/Context';
+} from '../../context/Overlay';
 
 import LanguageListOverlay from './LanguageListOverlay';
