+++ src/components/LanguageSelector.jsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
 
-import Context, { SET_LANGUAGE_OVERLAY_VISIBILITY } from '../context/Context';
+import Context, { SET_LANGUAGE_OVERLAY_VISIBILITY } from '../context/Overlay';
 
 import Icon from './Icon';
